Adds maxYear to snapshotsManager
Changes Timeline Slider to use maxYear

Changes Timeline Slider to use a custom track to show time progress instead of the slider track.

Bug is when sim is started to continue playing, the time progress starts where the thumb is and vegetation stat graph starts at the end, but shows the vegetation stat where thumb is.